refactor: Finish Quickstart tutorial
