+++ cognite/neat/rules/exporter/_rules2pydantic_models.py
