+++ docs/api/rules/exporters.md
@@ -1,9 +1,3 @@
-::: cognite.neat.rules.exporter.rules2excel.ExcelExporter
+# Exporters
 
-::: cognite.neat.rules.exporter.rules2dms.DataModel
-
-::: cognite.neat.rules.exporter.rules2ontology.Ontology
-
-::: cognite.neat.rules.exporter.rules2graphql.GraphQLSchema
-
-::: cognite.neat.rules.exporter.rules2pydantic_models.rules_to_pydantic_models
+::: cognite.neat.rules.exporter
