+++ mkdocs.yml
@@ -84,6 +84,9 @@ plugins:
       execute: false
       ignore_h1_titles: True
       include_source: True
+      remove_tag_config:
+        remove_input_tags:
+          - hide
   - glightbox:
       width: 100%
       height: auto
