+++ tests/rules/importer/test_json2tables.py
@@ -4,7 +4,7 @@ from tests import data
 
 def test_json2rules_power_grid() -> None:
     # Arrange
-    json_importer = importer.JSONImporter(data.POWER_GRID_JSON)
+    json_importer = importer.JSONImporter(data.POWER_GRID_JSON, "child-to-parent")
     expected_properties = {
         "GeographicalRegion": {"name"},
         "SubGraphicalRegion": {"name", "parent"},
